Oh, common! (#13)
* feat: fix scrollbar part 1

* feat: change button animation at settings panel

* feat: change animation for all buttons

* feat: change checkbox styles

* feat: first step to build logic in settigs panel's checkboxes(add autosave checkbox logic)

* feat: remove loginButton because ist dont need

* feat: fix bag in setting panel

* feat: remove FullPlayList component, because we have playlist comp, and delete second arg in DsongFn

* feat: formatting (use shift+alt+F) =))

* feat: index.ts rename => index.tsx

* feat: fix webpack problem

* feat: add modal to view client's avatar

* feat: add target _blank for open links in new page

* feat: fix userCard/playlist if we dont have data. And remove images from scripts file in src

* feat: remove logout button because its overcode

* fix(SettingsPanel): fix DW palylist id form autofill

* fix styles: fixed checkbox hover

+++ src/components/SongCard.tsx
@@ -1,21 +1,36 @@
-import React from "react"
-import PropTypes from "prop-types"
-const SongCard = ({ song, index, isDeletable, onDelete, isHidden }) => {
+import React, { FC } from 'react';
+import { Song } from '../interfaces/Song';
+
+interface ISongCard {
+    song: Song;
+    index: number;
+    isDeletable: boolean;
+    onDelete?: (index: number, value: boolean) => void;
+    isHidden?: boolean;
+}
+
+const SongCard: FC<ISongCard> = ({
+    song,
+    index,
+    isDeletable,
+    onDelete,
+    isHidden,
+}) => {
     const handleChange = () => {
         if (isDeletable) {
-            console.log("Deletable")
-            onDelete(index, true)
+            console.log('Deletable');
+            if (onDelete) onDelete(index, true);
         }
-    }
+    };
     return (
         <div
             className={`flex border-solid border-2 m-3 rounded-md ${
-                isHidden && "hidden"
+                isHidden && 'hidden'
             }`}
         >
             <div
                 className={`text-black p-2 mb-auto mt-auto inline ${
-                    index < 10 && "ml-2"
+                    index < 10 && 'ml-2'
                 }`}
             >
                 {index}
@@ -36,8 +51,8 @@ const SongCard = ({ song, index, isDeletable, onDelete, isHidden }) => {
                     {song.artists}
                 </div>
             </div>
-            <div
-                className={`text-white mr-5 my-auto ${!isDeletable && "hidden"}
+            <button
+                className={`text-white mr-5 my-auto ${!isDeletable && 'hidden'}
             border-solid border-2 m-3 rounded-md px-5 
             hover:bg-red-400 hover:pointer hover:border-transparent cursor-pointer
             transition
@@ -45,23 +60,9 @@ const SongCard = ({ song, index, isDeletable, onDelete, isHidden }) => {
                 onClick={handleChange}
             >
                 X
-            </div>
+            </button>
         </div>
-    )
-}
+    );
+};
 
-SongCard.defaultProps = {
-    onDelete: () => {},
-    isHidden: false,
-    isDeletable: false,
-}
-
-// SongCard.propTypes = {
-//     song: PropTypes.string,
-//     index: PropTypes.string,
-//     onDelete: PropTypes.func,
-//     isHidden: PropTypes.boolean,
-//     isDeletable: PropTypes.boolean,
-// }
-
-export default SongCard
+export default SongCard;
